Move bits.
